app/requirements.txt
click~=8.1.3
requests~=2.28.0
setuptools~=61.2.0
typing_extensions~=4.2.0
pillow~=9.2.0
numpy~=1.23.1
python-dotenv~=0.20.0
prettytable~=3.3.0